map retived form file and loaded, not yet tested

+++ src/Estrada.cpp
@@ -5,8 +5,8 @@ Estrada::Estrada(int id, Node* ini, Node *dest) {
 	this->id = id;
 	this->ini =ini;
 	this->dest = dest;
-	int x = dest->getX() - ini->getX();
-	int y = dest->getY() - ini->getY();
+	float x = dest->getX() - ini->getX();
+	float y = dest->getY() - ini->getY();
 	this->peso = sqrt((x*x) + (y*y));
 }
