+++ src/map.h
@@ -2,6 +2,14 @@
 #define MAP_H_
 
 #include <string>
+#include <stdio.h>
+#include <unistd.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <fcntl.h>
+
+
+
 #include "Estrada.h"
 #include "node.h"
 #include "graphviewer.h"
@@ -15,7 +23,7 @@ private:
 	std::vector<Estrada> estradas;
 	std::vector<Node> solucao;
 
-	void addNode (int id, int x, int y, Tipo tipo);
+	void addNode (int id, float x, float y, Tipo tipo);
 	void addEstrada (int id, int nodeIdInicio, int nodeIdDestino);
 public:
 	Map (string cidade);
