+++ src/node.h
@@ -18,21 +18,21 @@ class Node {
 private:
 		int id;
 		Tipo tipo;
-		int x, y;
+		float x, y;
 		std::vector<Estrada> estradas;
 		bool visited;
 
 		void addEstrada(Node *dest, double peso);
 		bool removeEdgeTo(Node *dest);
 public:
-		Node(int x, int y, int id, Tipo tipo);
+		Node(float x, float y, int id, Tipo tipo);
 
 		std::vector<Estrada> getEstradas();
 		void addEstrada (Estrada estrada);
 		int getId();
 		Tipo getTipo();
-		int getX();
-		int getY();
+		float getX();
+		float getY();
 		bool isVisited();
 		void setVisited();
 };
